A Very simple echo client. Code is for one client only and 2 basic commands, !say and !shutdown.
What it does is repeat anything that doesn't begin with an explamation point (!) When it echos back to the server it appends a ! at the start of the message to prevent bouncing between two echo bots.

This is to be used for testing, will write test scripts to stress test connecting and disconnecting at a later point.