1)  Large update in sequencer and the Killer thread, see below 2)  Made STACKLOG a macro that can be turned off to increase performance by reducing the number of calls to the logger, this was the most called function in the server. 3)  Minor code cleanups
About the change to the sequencer and killer thread:
Killer thread would acquire the clients mutex then get stuck waiting for a client socket to disconnect, deadlocking the entire server.

The old implementation kept an array of clients in memory, the killer thread changed the client in this array, which meant the clients mutex needed to be locked. Instead of keeping an array of clients in memory they are dynamically allocated, when they are disconnected they are removed from the clients list and added to the killer thread list. The only calls that use a mutex are called to add or remove items from the killer queue, this more concurrency is possible, and the killer thread cannot block the sequencer, nor will it block adding of clients to the kill queue.