Now supporting net performance graphs on client.
See https://github.com/RigsOfRods/rigs-of-rods/pull/3056 - `RoRnet::Header` was updated with extra timestamps to track network performance. The problem is, the server didn't transfer the header as-is, so the data got lost. I added code to transfer the headers as-is, but to avoid massive commit, I left all the server messages as-is, only changed processing messages from clients.

Changes:
- RoRnet updated to support new timestamps
- Messaging: added `SendMessageWithHeader()`
- Broadcaster - transformed `QueueMessage()` with separate values to `QueuePacket` with header.
- Sequencer - accept and broadcast the entire header. Added `sendStreamData()` helper for that.